Fix paths in generate_template.py

+++ src/pca_report_library/templates/generate_template.py
@@ -33,45 +33,28 @@ import pkg_resources
 
 from .._version import __version__
 
-MANUAL_DATA_TEMPLATE_FILE_NAME = "template-manualData_RVXXXX.json"
-MUSTACHE_FILE_NAME = "report.mustache"
+MANUAL_DATA_TEMPLATE = "template-manualData_RVXXXX.json"
+ASSETS_DIR = pkg_resources.resource_filename("pca_report_library", "assets")
+MUSTACHE_FILE = "report.mustache"
+
 ASSETS_DIR_DST = "assets"
-MUSTACHE_FILE = pkg_resources.resource_filename(
-    "pca_report_library",
-    "{asset_dir}/{mustache_filename}".format(
-        asset_dir=ASSETS_DIR_DST, mustache_filename=MUSTACHE_FILE_NAME
-    ),
-)
-MANUAL_DATA_TEMPLATE_FILE = pkg_resources.resource_filename(
-    "pca_report_library",
-    "{asset_dir}/{manual_data_filename}".format(
-        asset_dir=ASSETS_DIR_DST, manual_data_filename=MANUAL_DATA_TEMPLATE_FILE_NAME
-    ),
-)
-ASSETS_DIR_SRC = pkg_resources.resource_filename("pca_report_library", "assets")
 
 
 def manualData_template(work_dir):
     """Move manual data template to working dir."""
-    me = os.path.realpath(__file__)
-    my_dir = os.path.dirname(me)
-    file_src = os.path.join(my_dir, MANUAL_DATA_TEMPLATE_FILE)
-    file_dst = os.path.join(work_dir, MANUAL_DATA_TEMPLATE_FILE)
+    file_src = os.path.join(ASSETS_DIR, MANUAL_DATA_TEMPLATE)
+    file_dst = os.path.join(work_dir, MANUAL_DATA_TEMPLATE)
     shutil.copyfile(file_src, file_dst)
 
 
 def latex_template(work_dir):
     """Move latex template to working dir."""
-    me = os.path.realpath(__file__)
-    my_dir = os.path.dirname(me)
-
-    file_src = os.path.join(my_dir, MUSTACHE_FILE)
-    file_dst = os.path.join(work_dir, MUSTACHE_FILE_NAME)
+    file_src = os.path.join(ASSETS_DIR, MUSTACHE_FILE)
+    file_dst = os.path.join(work_dir, MUSTACHE_FILE)
     shutil.copyfile(file_src, file_dst)
     # copy static assets
-    dir_src = os.path.join(my_dir, ASSETS_DIR_SRC)
     dir_dst = os.path.join(work_dir, ASSETS_DIR_DST)
-    shutil.copytree(dir_src, dir_dst)
+    shutil.copytree(ASSETS_DIR, dir_dst)
 
 
 def main():
@@ -86,7 +69,7 @@ def main():
 
     if args["--manualData"]:
         manualData_template(temp_working_dir)
-        print("Copying {} to current dir....".format(MANUAL_DATA_TEMPLATE_FILE))
+        print("Copying {} to current dir....".format(MANUAL_DATA_TEMPLATE))
 
     if args["--LaTeX"]:
         latex_template(temp_working_dir)
@@ -94,7 +77,8 @@ def main():
         print("Copying Asset Folder for LaTeX report to current dir....")
 
     for i in os.listdir(temp_working_dir):
-        shutil.move(os.path.join(temp_working_dir, i), original_working_dir)
+        if not os.path.exists(os.path.join(temp_working_dir, i)):
+            shutil.move(os.path.join(temp_working_dir, i), original_working_dir)
 
     os.chdir(original_working_dir)
     shutil.rmtree(temp_working_dir)
